// src/ErrorDashboard/Client/src/api/types.gen.ts
export type OffendingPageResponseModel = {
    host: string;
    path: string;
    type: string;
    statusCode?: null | number;
    count: number;
    firstSeenUtc: string;
    lastSeenUtc: string;
    documentKey?: null | string;
    isUnpublished: boolean;
    hasReferrers: boolean;
};
// ...
export type PageSortField = 'Count' | 'LastSeen' | 'Path';
// ...
export type ReferrerResponseModel = {
    referrer?: null | string;
    count: number;
    lastSeenUtc: string;
};

export type ReferrersResponseModel = {
    items: Array<ReferrerResponseModel>;
    availableDays: number;
};
// ...
export type SortDirection = 'Ascending' | 'Descending';